requirements.txt
appnope=0.1.2=py38hecd8cb5_1001
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boto3=1.17.78=pyhd8ed1ab_0
botocore=1.20.78=pyhd8ed1ab_0
brotlipy=0.7.0=py38haf1e3a3_1000
bzip2=1.0.8=h1de35cc_0
ca-certificates=2021.5.30=h033912b_0
certifi=2021.5.30=py38hecd8cb5_0
cffi=1.14.5=py38h2125817_0
chardet=4.0.0=pypi_0
click=8.0.1=py38h50d1736_0
cryptography=3.4.7=py38h2fd3fbb_0
cycler=0.10.0=py38_0
cython=0.29.22=pypi_0
decorator=5.0.3=pyhd3eb1b0_0
ffmpeg=4.2.2=h97e5cf8_0
freetype=2.10.4=ha233b18_0
gettext=0.20.2=h7535e17_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.5=h91ad68e_1002
grad-cam=1.1.1=pypi_0
helper=2.5.0=pyh44b312d_0
icu=58.2=h0a44026_3
idna=2.10=pypi_0
importlib-resources=5.1.2=pypi_0
intel-openmp=2019.4=233
ipykernel=5.5.5=py38h6c79ece_0
ipython=7.18.1=py38h5ca1d4c_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jedi=0.18.0=py38hecd8cb5_1
jmespath=0.10.0=py_0
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=he5867d9_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py38hecd8cb5_0
kiwisolver=1.3.1=py38h23ab428_0
lame=3.100=h1de35cc_0
lcms2=2.11=h92f6f08_0
libcxx=10.0.0=1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libsodium=1.0.18=h1de35cc_0
libtiff=4.1.0=hcb84e12_1
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
libxml2=2.9.10=h7cdb67c_3
llvm-openmp=10.0.0=h28b9765_0
lz4-c=1.9.3=h23ab428_0
matplotlib=3.3.4=py38hecd8cb5_0
matplotlib-base=3.3.4=py38h8b3ea08_0
mkl=2019.4=233
mkl-service=2.3.0=py38h9ed2024_0
mkl_fft=1.3.0=py38ha059aab_0
mkl_random=1.1.1=py38h959d312_0
ncurses=6.2=h0a44026_1
nettle=3.4.1=h3018a27_0
ninja=1.10.2=py38hf7b0b51_0
numpy=1.19.2=py38h456fd55_0
numpy-base=1.19.2=py38hcfb5961_0
olefile=0.46=py_0
opencv-contrib-python=4.5.1.48=pypi_0
opencv-python=4.5.2.52=pypi_0
openh264=2.1.0=hd9629dc_0
openssl=1.1.1k=h9ed2024_0
packaging=20.9=pypi_0
pandas=1.2.3=py38hb2f4e1b_0
parso=0.8.2=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.0.0=py38h1a82f1a_0
pip=21.0.1=py38hecd8cb5_0
prompt-toolkit=3.0.17=pyh06a4308_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pycocotools=2.0.2=pypi_0
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pymongo=3.11.3=pypi_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py38_1
python=3.8.8=h88f2d9e_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python-dotenv=0.17.1=pyhd8ed1ab_0
python_abi=3.8=1_cp38
pytorch=1.8.0=py3.8_0
pytz=2021.1=pyhd3eb1b0_0
pyyaml=5.4.1=py38h9ed2024_1
pyzmq=20.0.0=py38h23ab428_1
readline=8.1=h9ed2024_0
requests=2.25.1=pypi_0
s3transfer=0.4.2=pyhd8ed1ab_0
scikit-learn=0.24.2=py38hb2f4e1b_0
scipy=1.6.2=py38h2515648_0
seaborn=0.11.1=pyhd3eb1b0_0
setuptools=52.0.0=py38hecd8cb5_0
simpledorff=0.0.2=pypi_0
six=1.15.0=py38hecd8cb5_0
sqlite=3.35.3=hce871da_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hb0a8c7a_0
torchcam=0.2.0=0
torchray=1.0.0.2=pypi_0
torchvision=0.9.0=py38_cpu
tornado=6.1=py38h9ed2024_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=pyha847dfd_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
wheel=0.36.2=pyhd3eb1b0_0
x264=1!157.20191217=h1de35cc_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zlib=1.2.11=h1de35cc_3
zstd=1.4.9=h322a384_0
